feat: show network status on title
