+++ src/components/Title/index.js
@@ -1,6 +1,23 @@
 import React from "react";
-import styled from "styled-components";
+import styled, { keyframes } from "styled-components";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faWifi } from "@fortawesome/free-solid-svg-icons";
 import { baseColors } from "../../theme";
+import { useOnline } from "../../containers/Online";
+
+const fade = keyframes`
+0% {
+  opacity: 0.2;
+}
+
+50% {
+  opacity: 1;
+}
+
+100% {
+  opacity: 0.2;
+}
+`;
 
 const StyledTitle = styled.div`
   background: ${baseColors.background};
@@ -8,6 +25,7 @@ const StyledTitle = styled.div`
   border-radius: 0 0 5px 5px;
   align-self: center;
   transition: all 1s ease;
+  text-align: center;
   z-index: 10;
 
   box-shadow: ${({ shadow }) =>
@@ -33,17 +51,34 @@ const StyledTitle = styled.div`
     height: 80px;
   }
 
-  > * {
+  > * > * {
+    display: inline-block;
     margin: 0;
     padding: 0.2em;
     font-size: 1.5em;
     color: ${baseColors.heading};
     text-align: center;
   }
+
+  .network-fade {
+    color: ${baseColors.heading};
+    animation: ${fade} 2s ease-in infinite;
+  }
 `;
 
-export function Title({ children, ...props }) {
-  return <StyledTitle {...props}>{children}</StyledTitle>;
+const Network = () => (
+  <FontAwesomeIcon title="No Internet" icon={faWifi} className="network-fade" />
+);
+
+export function Title({ network, children, ...props }) {
+  const online = useOnline();
+
+  return (
+    <StyledTitle {...props}>
+      <span>{children}</span>
+      <span>{!online && <Network />}</span>
+    </StyledTitle>
+  );
 }
 
 export default React.memo(Title);
